refactor(SpentTable): move component to its own directory
